one possible solution
You can correct the initial formatting in the samplesudoku2.txt and samplesudoku4.txt files (delete the space between the first and second lines, and add it back). This code works with samplesudoku2.txt, but not samplesudoku4.txt, because it's not smart enough to solve where there are 53 empty spaces initially (this may also have something to do with where the spaces are in the board). Perhaps we can combine and refactor now?