Create updated_Borders-sudoku.cpp
I identified 4 "cell" answers that should be obvious in samplesudoku4.txt, and created a way to fill in those blanks, but my solution only hits 3 out of the 4 cells.

The one left to address is temp[5][4], which is 6 because...
- temp[3][1] is 6, so 6 can't go in row 3, and 
- temp[5][8] (the only other possible place to put a 6 in row 5) cannot be 6 because temp[0][8] is 6, meaning in that 3x3 grid, 6 must go somewhere in row 4, and 
- thus in the 3x3 grid where temp[5][4] is located, 6 also cannot be in row 4.

I know what's left to be done on this, but I have *got* to stop so I can study for the final. I can't look at it anymore; my eyes are starting to cross. I surrender my code at this point--add what's missing, clean it up, encapsulate it, etc. I have to walk away from it until after the exam. Sorry!